requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
cloudinary==1.25.0
cycler==0.10.0
Cython==0.29.23
Django==2.1.5

gunicorn==20.1.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.4.2
numpy==1.20.3
django-heroku==0.3.1
opencv-python==4.5.2.52
pandas==1.2.4
Pillow==8.2.0
psycopg2==2.8.6
pyparsing==2.4.7
pytesseract==0.3.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
six==1.16.0
sqlparse==0.4.1
tqdm==4.60.0
whitenoise==5.2.0
typing-extensions==3.10.0.0
